Update links & blurb
I have updated the logo so it links to the landing page from the recipe pages.
Have updated the inspire me button to a link to the recipe on index.html, and adjusted styling.
Updated blurb on landing page to something a little more relevant to a baking website.
Changed thumbnail on landing page to the roulade.  Commented out the a tag but not removed, with intention of reinstating in future.